Added user input checking
For the feature that creates folders and shares for user accounts, the base directory shouldn't have trailing slashes and the folder name should not have leading or trailing slashes. If they do, those are removed now. Also, if the Create Folder For Users radio button is set to true, then base directory and folder name must be completed. If either of these two are not completed, automatically set Create Folder For Users to false but store the input given, if any.