Started work on Jane Sentinel
Jane Sentinel is intended to be a helper service for Jane's mission. It is a BASH script that runs via a cron schedule as root. It checks that the Jane Engine is running, and possibly other services too. If anything is wrong, it attempts to fix it. If all fix attempts fails, it should wait for 30 minutes and then do a system reboot as a last effort. At the moment, the feature is incomplete and untested.