Added log archiver and finished sentinel.
The log archiver by default moves the current JaneEngine.log file to a file with a date and time stamp, and creates a new log file, sets ownership and permissions of this file, and notes in the new log that the old log has been moved. The janeSentinel program should now attempt to restart all services if the JaneEngine.php is not detected as running. On failure of critical items, it should wait 30 minutes (default) before attempting to reboot.